pyproject.toml
[project]
name = "secret-santa-cli"
version = "0.1.0"
description = "A Secret Santa game CLI tool."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.0",
    "jinja2>=3.1.4",
    "pydantic-settings>=2.7.0",
    "pydantic>=2.10.3",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = ["ipython>=8.30.0", "pytest>=8.3.4", "ruff>=0.8.1"]
